Create syn-stage00.sh
This script is written in Bash, a Unix shell and command-language. It is intended to be used to set up a system, it sets keyboard layout, enables Network Time Protocol (NTP) and starts DHCP service.

This script starts by setting the keyboard layout to "uk" using the loadkeys command, then it enables NTP using the timedatectl command and finally it starts the DHCP service using systemctl command.

The script then gives a warning message that it has the potential to cause damage to the system if not executed properly, and that any consequences resulting from the improper use of this script will be solely the user's responsibility.

The script then sets partition sizes and defines the disk that will be wiped, it also prompts the user to confirm the disk info before proceeding.

It then uses parted command to create GPT label and partitions and then creates file systems, mount partitions and creates boot directory.

It then proceeds to format the partitions, mount them and copy the data to the new system.

It also gives a warning that this script has not implemented advanced features such as non-standard file systems such as xfs, zfs, or btrfs for block-device backups and encryption for both the live and persistent environments.

It is important to note that this script is not complete and requires user input and additional commands to properly function.

It is also important to note that this script is irreversible and can cause serious data loss if not used properly. It is advised to use this script only if you are familiar with the commands used in it and have a clear understanding of what it does.